+++ src/app/document/core/pdf_loader/pdfminer_loader.py
