+++ src/app/document/core/pdf_loader/pdfplumber_loader.py
