+++ src/app/document/func/text_helper.py
@@ -5,9 +5,10 @@ import logging
 from typing import List, Dict, Any, Optional
 from datetime import datetime
 from src.app.document.model import DocumentChunk
-from src.app.core.config import settings
+from src.app.document.core.config import settings
+from src.common.utils.logger import set_logger
 
-logger = logging.getLogger(__name__)
+logger = set_logger(__name__)
 
 
 class TextHelper:
