+++ src/app/document/service.py
@@ -2,7 +2,7 @@ import time
 
 from fastapi import UploadFile
 
-from src.app.core.pdf_loader.factory import PDFLoaderFactory
+from src.app.document.core.pdf_loader.factory import PDFLoaderFactory
 from src.app.document.func.document import Document
 from src.app.document.func.document_loader import DocumentLoader
 from src.app.document.func.vector import Vector
